Kill overlap stubs in InputData, set in Settings
That is, the way of killing overlap stubs (none, pairFinder,
truePairFinder) is set via a Python script so that I do not need to
recompile upon changing options.

Overlap stubs are now killed in InputData instead of the TMTrackProducer
so that the effect of this can be histogrammed by histograms of input
data.